third_party/blink/web_tests/external/wpt/css/css-anchor-position/last-successful-change-try-rule-expected.txt

This is a testharness.js-based test.
[FAIL] No successful position, keep --try
  assert_equals: expected 200 but got 250
[FAIL] No successful position, last successful invalidated by @position-try change
  assert_equals: expected 0 but got -50
Harness: the test ran to completion.
